+++ NEWS.md
@@ -2,6 +2,7 @@
 - Daily substeps available for base flows
 - Aquifer (water table depth) effects on soil local water balance can be disabled
 - Aquifer water balance revised
+- Negative aquifer correction introduced
 
 # medfateland 2.7.1
 - Subwatersheds returned as part of spatial output in *_land simulations